+++ packages/index.ts
@@ -1,2 +1,1 @@
-export { VisualRegression } from './service';
-export { Matcher } from './service/matchers/base-matcher';
+export * from './public-api';
