+++ packages/public-api.ts
@@ -1,0 +1,3 @@
+export { VisualRegression } from './service';
+export { defineMatcher } from './service/core/define-matcher';
+export { Matcher } from './service/core/base-matcher';
